ms-cli: hashlock's --separator binds to the shared parser and --group-size takes encode's type (post-impl review I-2, N-1)
REPRODUCED FIRST, on the shipped binary:

  $ ms hashlock --in X.txt --separator q 2>&1 >/dev/null | grep 'preimage (ms1)'
  preimage (ms1):  ms10hqashsqq0p7jaqf9gsjqjpkjvqll2l2q74w8aq388xgqqzlewqp73scqptwxgqtjugsqpvs8tqklufgq89hqj
  $ ms encode --hex 000102...0f --separator q --allow-argv-secret
  error: invalid value 'q' for '--separator <SEPARATOR>': invalid separator "q"; expected `space` (or the literal " ")

`q` is in the codex32 charset, so `strip_display_separators` will not remove it
on read-back: a plate cut from that card is a 90-character string `ms` itself
refuses. `--separator -` was accepted here and refused by encode/split too.
N-1, same shape: `--group-size 256` -> `256 is not in 0..=255` on hashlock,
accepted on encode.

Provenance: the plan carried `#[arg(long, default_value_t = ' ')]` verbatim
(IMPLEMENTATION_PLAN_ms_hashlock_H1.md:2271-2273), so this is a plan omission
reproduced, not an implementer deviation -- fixed here rather than inherited.

FIX: `#[arg(long, default_value = "space", value_parser =
crate::format::parse_separator)]` and `group_size: u16`, character for
character what `cmd/encode.rs:85-90` and `cmd/split.rs` already carry.
`parse_separator`'s doc said it "serves both `ms encode` and `ms split`"; it
serves three verbs now.

AFTER:
  $ ms hashlock --in X.txt --separator q      exit=64
  error: invalid value 'q' for '--separator <SEPARATOR>': invalid separator "q"; expected `space` (or the literal " ")
  $ ms hashlock --in X.txt --separator -       exit=64
  error: invalid value '-' ...: separator "-" is no longer offered: `ms` emits whitespace grouping only. ...
  $ ms hashlock --in X.txt --group-size 256    exit=0

NEW TESTS, both proven to fail on the revert:
`separator_is_refused_exactly_as_encode_refuses_it` compares the parser's own
sentence between `hashlock` and `encode` for q, - and , (not merely "both
refused") and keeps `space` and " " working;
`group_size_accepts_what_encode_accepts` pins 256.
Reverting both attributes:
  test group_size_accepts_what_encode_accepts ... FAILED
  test separator_is_refused_exactly_as_encode_refuses_it ... FAILED
  assertion `left == right` failed: --separator "q" must be a usage error ...
    left: Some(0)
   right: Some(64)
Reverted; hashlock_outputs -> 11 passed.

+++ crates/ms-cli/src/cmd/hashlock.rs
@@ -73,10 +73,17 @@ pub struct HashlockArgs {
     #[arg(long)]
     pub no_engraving_card: bool,
     /// Group the ms1 on the card every N characters (0 = no grouping).
+    /// `u16` and the same default as `ms encode` / `ms split`, so the same
+    /// value is accepted by every verb that renders a grouped ms1 (review N-1).
     #[arg(long, default_value_t = 5)]
-    pub group_size: u8,
-    /// Group separator on the card.
-    #[arg(long, default_value_t = ' ')]
+    pub group_size: u16,
+    /// Separator: space|hyphen|comma (keyword) or the literal " "|-|, . SPEC §5.
+    ///
+    /// BOUND TO THE SHARED PARSER, like `ms encode` and `ms split`. Unbound, a
+    /// separator inside the codex32 charset (`--separator q`) produced a card
+    /// whose "plate string" `strip_display_separators` cannot clean up, so the
+    /// engraved 90-character result is one `ms` itself refuses (review I-2).
+    #[arg(long, default_value = "space", value_parser = crate::format::parse_separator)]
     pub separator: char,
     /// Admit a secret on argv (see `ms encode --help`).
     #[arg(long)]
